Port macos build fixes from master

+++ dbms/src/Storages/MergeTree/DiskSpaceMonitor.h
@@ -96,17 +96,17 @@ public:
 
     using ReservationPtr = std::unique_ptr<Reservation>;
 
-    static size_t getUnreservedFreeSpace(const std::string & path)
+    static UInt64 getUnreservedFreeSpace(const std::string & path)
     {
         struct statvfs fs;
 
         if (statvfs(path.c_str(), &fs) != 0)
             throwFromErrno("Could not calculate available disk space (statvfs)", ErrorCodes::CANNOT_STATVFS);
 
-        size_t res = fs.f_bfree * fs.f_bsize;
+        UInt64 res = fs.f_bfree * fs.f_bsize;
 
         /// Heuristic by Michael Kolupaev: reserve 30 MB more, because statvfs shows few megabytes more space than df.
-        res -= std::min(res, static_cast<size_t>(30 * (1ul << 20)));
+        res -= std::min(res, static_cast<UInt64>(30 * (1ul << 20)));
 
         std::lock_guard<std::mutex> lock(mutex);
 
@@ -118,7 +118,7 @@ public:
         return res;
     }
 
-    static size_t getReservedSpace()
+    static UInt64 getReservedSpace()
     {
         std::lock_guard<std::mutex> lock(mutex);
         return reserved_bytes;
@@ -131,9 +131,9 @@ public:
     }
 
     /// If not enough (approximately) space, throw an exception.
-    static ReservationPtr reserve(const std::string & path, size_t size)
+    static ReservationPtr reserve(const std::string & path, UInt64 size)
     {
-        size_t free_bytes = getUnreservedFreeSpace(path);
+        UInt64 free_bytes = getUnreservedFreeSpace(path);
         if (free_bytes < size)
             throw Exception("Not enough free disk space to reserve: " + formatReadableSizeWithBinarySuffix(free_bytes) + " available, "
                 + formatReadableSizeWithBinarySuffix(size) + " requested", ErrorCodes::NOT_ENOUGH_SPACE);
@@ -141,7 +141,7 @@ public:
     }
 
 private:
-    static size_t reserved_bytes;
+    static UInt64 reserved_bytes;
     static size_t reservation_count;
     static std::mutex mutex;
 };
