+++ dbms/src/Storages/MergeTree/MergeTreeDataMergerMutator.h
@@ -45,12 +45,12 @@ public:
     /** Get maximum total size of parts to do merge, at current moment of time.
       * It depends on number of free threads in background_pool and amount of free space in disk.
       */
-    size_t getMaxSourcePartsSize();
+    UInt64 getMaxSourcePartsSize();
 
     /** For explicitly passed size of pool and number of used tasks.
       * This method could be used to calculate threshold depending on number of tasks in replication queue.
       */
-    size_t getMaxSourcePartsSize(size_t pool_size, size_t pool_used);
+    UInt64 getMaxSourcePartsSize(size_t pool_size, size_t pool_used);
 
     /** Selects which parts to merge. Uses a lot of heuristics.
       *
